requirements.txt
opencv-python==4.8.0.74
imgaug==0.4.0
pandas==1.5.3
torchvision==0.15.2
Pillow==8.4.0
numpy==1.25.2